pyproject.toml
[tool.poetry]
name = "git-tools"
version = "0.1.0"
description = "a smattering of git-related tools"
authors = ["Scott A. Idler <scott.a.idler@gmail.com>"]
license = "mit"
readme = "README.md"
packages = [{ include = "git_tools" }]

[tool.poetry.scripts]
clone = "git_tools.clone:main"
clone-lite ="git_tools.clone_lite:main"
filter-ref = "git_tools.filter_ref:main"
ls-github-repos = "git_tools.ls_github_repos:main"
stale-branches = "git_tools.stale_branches:main"

[tool.poetry.dependencies]
python = "^3.10"
ruamel-yaml = {version = "^0.18.6", optional = true}
requests = {version = "^2.31.0", optional = true}

[tool.poetry.extras]
ls-github-repos = ["requests"]
stale-branches = ["ruamel-yaml"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
